% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{auc}
\alias{auc}
\title{area under the curve}
\usage{
auc(data, pos = TRUE)
}
\arguments{
\item{data}{: a numeric vector}

\item{pos}{: a logical value. if TRUE (default) AUC is calculated on positive values, if FALSE on negative values.}
}
\value{
auc : area under the curve

max/min : maxiaml value (or minimal if pos=FALSE)
}
\description{
take a numeric vector and return the area under the curve (AUC)
}

