% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biosensor.R
\name{calibration}
\alias{calibration}
\title{Sensor calibration}
\usage{
calibration(volt, mol, order = 1)
}
\arguments{
\item{volt}{: a numeric vector of the biosensor voltage}

\item{mol}{: a numeric vector of the molecule conentration in the medium (x and y must have the same length)}

\item{order}{: a numeric value for the polynomial degree. default is one.}
}
\value{
Coef : coeficients in ascending order (i.e. C0, C1, C2, \ldots , Cn )

R2 : goodness of fit
}
\description{
Fit an n\eqn{^{th}}{th} degree polynom
\deqn{(y = C_nX^n + C_{n-1}X^{n-1} + \ldots + C_1X + C_0)}{
(y = CnX^n + Cn-1X^n-1 + \ldots + C1X + C0)}
to vectors x=volt and y=concentration
}

