% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMRGluc.R
\name{CMRGluc.calc}
\alias{CMRGluc.calc}
\title{CMRGluc calculation}
\usage{
CMRGluc.calc(Gbrain, Vd = 0.77, Kt = 13.4, Tmax = 1.35, Gplasma = 6)
}
\arguments{
\item{Gbrain}{: a vector/numeric value of extracellular glucose concentration in mmol/L, time decay}

\item{Vd}{: glucose brain space diffusion (default is 0.77 ml/g)}

\item{Kt}{: glucose apparent maximal transport rate (default is 13.4 mmol/L)}

\item{Tmax}{: the apparent maximal transport rate (default is 1.35 micormol/g/min)}

\item{Gplasma}{: plasma glucose concentration (default=6mmol/L)}
}
\value{
CMRGlucose : numeric value of Cerebral metabolic rate of glucose in micromol/g/min
}
\description{
take extracellular glucose concentration and return brain metabolic rate of glucose, using a reversible Michaelis-Menten model equation:
}
\details{
\deqn{ G_{brain}=V_d \frac{\left(\frac{T_{max}}{CMR_{gluc}}-1\right)\times G_{plasma} - K_t}
{\frac{T_{max}}{CMR_{gluc}} +1} }{
Gbrain=Vd x ( ((Tmax)/(CMRgluc)-1) * Gplasma - Kt) / ((Tmax)/(CMRgluc) +1) }
}
\references{
Du et al JCBFM 2012 32(9)

Gruetter et al J Neurochem 1998 70(1)
}

