% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMRO2.R
\name{CMRO2.calc}
\alias{CMRO2.calc}
\title{CMRO2 calculation}
\usage{
CMRO2.calc(LDF, TPO2, P50 = 36, h = 2.7, Ca = 8, L = 4.03,
  cbfbase = 53)
}
\arguments{
\item{LDF}{: a vector/numeric value of LDF (percentage of baseline, i.e: basal value is 100 \% )}

\item{TPO2}{: a vector/numeric value of brain oxygnen pressure (mmHg)}

\item{P50}{: half-saturation tension of hemoglobine (default is 36 mmHg)}

\item{h}{: hill's coeficient}

\item{Ca}{: oxygen arterial concentration (default is 8 micromol/ml)}

\item{L}{: effective diffusion coefficient of oxygen in brain tissue, default is 4.03 micomol/100g/mmHg, but one should use the L.calc function to calculate it from their data.}

\item{cbfbase}{: basal expected value of CBF (default is 53 ml/100g/min wich was used to calculate L) \cr
LBF and TPO2 must be the same length}
}
\value{
CMRO2 : vector/numeric value of Cerebral Metabolic Rate of Oxygen in micromol/100g/min
}
\description{
take tissue oxygen pressure (tPO2) and cerebral blood flow (relative value, e.g. laser doopler lowmetry BPU)
}
\details{
\deqn{ PbtO_{2} = P_{50} \cdotp \sqrt[\scriptstyle h]{\frac{2 \cdotp C_{a} \cdotp CBF}{CMRO_{2}}-1} - \frac{CMRO_{2}}{2 \cdotp L} }{
PbtO2 = P50 x (2 * Ca * CBF/CMRO2-1)^(1/h) - CMRO2/(2*L)}
}
\references{
Gjedde et al JCBFM (2005) 25(9), 1183

Piilgaard et al JCBFM (2009) 29, 1517
}

