% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biosensor.R
\name{correction.Temp}
\alias{correction.Temp}
\title{Temperature correction}
\usage{
correction.Temp(x, enz, temp.calib = 25, temp.exp = 37)
}
\arguments{
\item{x}{: a numeric vector of the biosensor voltage}

\item{enz}{: enzyme on the biosensor i.e. "glucose", "lactate", "glutamate, "daao"}

\item{temp.calib}{: temperature of the medium where sensor has been calibrated
the default is 25\eqn{^o}{o}C (i.e. room temperature).}

\item{temp.exp}{: temperature of the medium during experiment
the default is 37\eqn{^o}{o}C (i.e. animal central temperature)}
}
\value{
volt.temp.cor : a vector of corected x values for temperature
}
\description{
Biosensor enzymatic reaction, that underpin amperometric measures, has a sigmoid relation to temperature :
\deqn{m(x,P)=\frac{P_1+(P_2-P_1)}{(1+\exp((P_3-x)/P_4))}}{
m(x,P)=P1+(P2-P1)/(1+exp((P3-x)/P4))}
parameters are different for each enzyme and has been measured in vitro.\cr
This function correct biosensor signal depending on temperature conditions during calibration and experimentation
}
\references{
balanca et al 2015
}

