% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/biosensor.R
\name{correction.TPO2}
\alias{correction.TPO2}
\title{Oxygen Tension correction}
\usage{
correction.TPO2(x, enz, TPO2 = 28)
}
\arguments{
\item{x}{: a numeric vector of the biosensor voltage}

\item{enz}{: enzyme on the biosensor i.e. "glucose", "lactate", "glutamate, "daao"}

\item{TPO2}{: oxygene tension in the medium during the experiment. the default is 30mmHg measured in anesthetized rat brain.}
}
\value{
volt.O2.cor : a vector of corected x values for TPO2
}
\description{
Biosensor enzymatic reaction, that underpin amperometric measures, has an asymptotic relation to oxygen tention in the medium :
\deqn{m(PO_2,P) = P_1 + (P_2 - P_1)\times \exp(-\exp(P_3)PO_2)}{
m(PO2,P) = P1 + (P2 - P1)*exp(-exp(P3)PO2)}
parameters are different for each enzyme and has been measured in vitro\cr
This function correct biosensor signal depending on PO2 conditions during calibration and experimentation
}
\references{
balanca et al 2015
}

