% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Index.R
\name{IO2.calc}
\alias{IO2.calc}
\title{Oxidative index calculation}
\usage{
IO2.calc(CMRO2, LDF, cbfbasal = 53)
}
\arguments{
\item{CMRO2}{: vector/numeric value of Cerebral metabolic rate of oxygen in micromol/100g/min}

\item{LDF}{:  a vector/numeric value of LDF (percentage from baseline, baseline is 100 \%)}

\item{cbfbasal}{: basale expected value of CBF from the litterature (default is 53 ml/100g/min wich was used to calculate L)}
}
\value{
IO2 : oxidative index, reflect the degree of flow metabolism coupling
}
\description{
take CMRO2 and LDF to give an oxidative index
}
\details{
\deqn{ IO2 = CMRO2/(cbfbasal*LDF) }
}
\references{
Gjedde et al JCBFM (2000) 20(4), 747
}

