% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CMRO2.R
\name{L.calc}
\alias{L.calc}
\title{Calculate the effective diffusion coefficient of oxygen in brain tissue, L.}
\usage{
L.calc(CMR = 219, TPO2, P50 = 36, h = 2.7, Ca = 8, cbf = 53)
}
\arguments{
\item{CMR}{: Cerebral Metabolic Rate of Oxygen, default is 219 micro mol/100}

\item{TPO2}{: numeric value of brain oxygnen pressure (mmHg)}

\item{P50}{: half-saturation tension of hemoglobine (default is 36 mmHg)}

\item{h}{: hill's coeficient}

\item{Ca}{: oxygen arterial concentration (default is 8 micromol/ml)}

\item{cbf}{: expected value of CBF (default is 53 ml/100g/min wich was used to calculate L}
}
\value{
L : numeric value of the effective diffusion coefficient of oxygen in brain tissue micomol/100g/mmHg
}
\description{
take CMRO2 and cerebral blood flow (CBF) to calculate the effective diffusion coefficient of oxygen in brain tissue (L)
}
\details{
\deqn{PbtO_{2} = P_{50} \cdotp \sqrt[\scriptstyle h]{\frac{2 \cdotp C_{a} \cdotp CBF}{CMRO_{2}}-1} - \frac{CMRO_{2}}{2 \cdotp L}}{
PbtO2 = P50 x (2 * Ca * CBF/CMRO2-1)^(1/h) - CMRO2/(2*L)}
}
\references{
Gjedde et al JCBFM (2005) 25(9), 1183

Piilgaard et al JCBFM (2009) 29, 1517
}

