% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{noise.na}
\alias{noise.na}
\title{Remove artifacts from biosensor signal, based on data's standar deviation}
\usage{
noise.na(data, z = 20, width = 30)
}
\arguments{
\item{data}{: a numeric vector}

\item{z}{: a numeric value. Number of SD  over which values should be exculded}

\item{width}{: size of the window used to roll SD over data (see roll.funct)}
}
\value{
a vector with NA remplacing exculded values
}
\description{
Remove artifacts from biosensor signal, based on data's standar deviation
}

