% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Index.R
\name{OGI.calc}
\alias{OGI.calc}
\title{Oxygene Glucose index (OGI)}
\usage{
OGI.calc(CMRO2, CMRGluc)
}
\arguments{
\item{CMRO2}{: vector (numeric value) of Cerebral metabolic rate of oxygen in micromol/100g/min}

\item{CMRGluc}{:  a vector (numeric value) of CMRGlucose in micromol/100g/min}
}
\value{
OGI
}
\description{
take CMRO2 and CMRGlucose to give an OGI
}
\details{
\deqn{ OGI = CMRO2/CMRGlucose }
}

