% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{polyval}
\alias{polyval}
\title{polynomial evaluation}
\usage{
polyval(x, coef)
}
\arguments{
\item{x}{: a numeric vector}

\item{coef}{: a n dimention vector corresponding to the polynom coefficient (ascending order, i.e. C0, C1, C2, ... , Cn )}
}
\value{
y
}
\description{
evaluation a n\eqn{^{th}}{th} degree polynom
\deqn{(y = C_nX^n + C_{n-1}X^{n-1} + \ldots + C_1X + C_0)}{(y = CnX^n + Cn-1X^n-1 + \ldots + C1X + C0)}
at given values of x
}

